requirements.txt
streamlit==1.27.2
groq
python-dotenv
